Add files via upload
this app is like in first installation and run a login window appears it takes a password among three which is already defined( i.e. GANGA , JAMUNA ,SARASWATI ), then it takes to the welcome page in top we  can see (HELLO 'password'), and some photos  with e-mail id , name and phone number arranged in a scrollview.